Added a click event to the naviconclickable view
When the view (the icon itself) is clicked on the page, the side navigation menu is expanded, the title text and page content is moved, and the navoverlay is shown, all by applying the appropriate classes via jquery. Conversely, if the side navigation menu is already shown, these changes are reversed.