Created the left-labeled-input component
Inputs are often listed below one another, and a similar format will be required for the add publications page. When inputs are listed below one another it is more readible to have them labeled to the left than to use hint text or top labels. This component provides an easily reusable template for inputs that are labeled to the left and arranged below one another.