Added the page class to the app scss
Ember creates extra divs in which your content is displayed when it is compiled, and since css flexing requires properties to be set in the direct parent element of the element that needs to be flexed, these styling rules could not be applied to the body element. Page content will now be wrapped in a page div.